Add all charts to Peity

+++ src/Blazorius.Charts.PeityBar/Models/BarChartOptions.cs
@@ -2,12 +2,12 @@
 {
     public class BarChartOptions : OptionsBase
     {
-        private int? max;
-        private int? min;
+        private double? max;
+        private double? min;
         private float? padding;
 
-        public int? Max { get => max; set => SetProperty(ref max, value); }
-        public int? Min { get => min; set => SetProperty(ref min, value); }
+        public double? Max { get => max; set => SetProperty(ref max, value); }
+        public double? Min { get => min; set => SetProperty(ref min, value); }
         public float? Padding { get => padding; set => SetProperty(ref padding, value); }
     }
 }
